Disable examples when RFM69_BUILD_EXAMPLES is off
Before, targets were created for the examples but they were not added to the
'all' target. However, when adding this library to a project, it is (1)
annoying to have these example targets, and (2) potentially problematic that
the examples pull-in the UartStdioAvr library because the main project may also
use that leading to duplicate CMake targets.