Do not rely on the default register values
Since the RFM69 module is not necessarily reset when the MCU is reset (e.g., it
is not on the Moteino) relying on the default register values is a bug. This
commit fixes the initialization routine to set all necessary registers even if
the used values are the default ones.